[refactor] Move high contrast `modify` overrides to separate hook in anticipation of new theme tokens
- there'll be a lot more, so let's pull it out to its own file and keep the main provider fairly high level in terms of high contrast mode logic
